server/pyproject.toml
[tool.poetry]
name = "django-nuxt-starter"
version = "0.1.0"
description = "Starter setup for django nuxt projects"
authors = ["moses@solframe.in"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
django-rest-framework = "^0.1.0"
psycopg2-binary = "^2.9.9"
django-cors-headers = "^4.4.0"
django-split-settings = "^1.3.2"
django-extensions = "^3.2.3"
setuptools = "^71.0.3"
pydantic = "^2.8.2"
celery = "^5.4.0"
drf-yasg = "^1.21.7"
djangorestframework-simplejwt = "^5.3.1"
gunicorn = "^22.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
